fix: disable layout animation in viewpager fragments (#720)
This fixes a bug where the viewpager stops mid-animation between two pages when opening some images.

A side effect of this change is that the extended details bubble will no longer animate when entering/exiting full-screen. The root cause should be fixed instead, but I do not have time to investigate it.